+++ experiments/lcwe/lcwe_model.c
@@ -1,5 +1,5 @@
 // Copyright 2026 Feralthedogg
-// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: LicenseRef-LLAM-Commercial-Reciprocity-1.0
 
 #include "lcwe_model_internal.h"