+++ scripts/docker_verify_linux.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 # Copyright 2026 Feralthedogg
-# SPDX-License-Identifier: Apache-2.0
+# SPDX-License-Identifier: LicenseRef-LLAM-Commercial-Reciprocity-1.0
 
 set -eu