+++ tests/test_shared_load.c
@@ -5,17 +5,7 @@
  * @copyright Copyright 2026 Feralthedogg
  *
  * @par License
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: LicenseRef-LLAM-Commercial-Reciprocity-1.0
  */
 
 #include "llam/runtime.h"
@@ -34,7 +24,7 @@ typedef int (*connect_fn)(llam_fd_t fd, const struct sockaddr *addr, socklen_t a
 #if defined(__APPLE__) && defined(__MACH__)
 #define LLAM_TEST_DEFAULT_SHARED_PATH "./libllam_runtime.2.dylib"
 #elif defined(__linux__)
-#define LLAM_TEST_DEFAULT_SHARED_PATH "./libllam_runtime.so.2.2.1"
+#define LLAM_TEST_DEFAULT_SHARED_PATH "./libllam_runtime.so.3.0.0"
 #else
 #define LLAM_TEST_DEFAULT_SHARED_PATH "./libllam_runtime.so"
 #endif